rename: RequestFileStrategy, ResponseFileStrategy 디렉토리 service에서 strategy로 이동
