+++ src/main/java/com/soda/article/strategy/ArticleFileStrategy.java
@@ -1,0 +1,4 @@
+package com.soda.article.strategy;
+
+public class ArticleFileStrategy {
+}
